+++ src/addin/client-interfaces/addin-client-callbacks.ts
@@ -16,12 +16,12 @@ export interface AddinClientCallbacks {
   buttonClick?: () => void;
 
   /**
-   * Callback raised for flyout add-ins indicating that the next button was clicked.
+   * Callback raised for tile add-ins indicating that the flyout's next button was clicked.
    */
   flyoutNextClick?: () => void;
 
   /**
-   * Callback raised for flyout add-ins indicating that the previous button was clicked.
+   * Callback raised for tile add-ins indicating that the flyout's previous button was clicked.
    */
   flyoutPreviousClick?: () => void;